Refactored redirections from apiServer to server-loader
Not tested
